Remove obsolete common fixtures file
